refactor(pkg): rename scrape to distro
