+++ pkg/distro/distro_centos.go
@@ -1,4 +1,4 @@
-package scrape
+package distro
 
 import (
 	"fmt"
