+++ pkg/distro/error.go
@@ -1,4 +1,4 @@
-package scrape
+package distro
 
 import "errors"
